microservicio_clientes_fastapi/requirements.txt
fastapi>=0.110.0
uvicorn>=0.28.0
sqlalchemy>=2.0.34
pydantic>=2.8.0
pydantic-settings>=2.1.0
python-dotenv>=1.1.0
alembic>=1.13.1
aiosqlite>=0.19.0
email-validator>=2.0.0